+++ CHANGELOG.md
@@ -1,3 +1,27 @@
+# v0.0.12 (Wed Oct 14 2020)
+
+#### 🐾 Patch
+
+- Dependent type issues [#19](https://github.com/gisat-panther/be-core/pull/19) ([@nenadalm](https://github.com/nenadalm))
+
+#### Authors: 1
+
+- Miloslav Nenadál ([@nenadalm](https://github.com/nenadalm))
+
+---
+
+# v0.0.11 (Wed Oct 14 2020)
+
+#### 🐾 Patch
+
+- Permission hashes [#18](https://github.com/gisat-panther/be-core/pull/18) ([@nenadalm](https://github.com/nenadalm))
+
+#### Authors: 1
+
+- Miloslav Nenadál ([@nenadalm](https://github.com/nenadalm))
+
+---
+
 # v0.0.10 (Fri Oct 09 2020)
 
 #### 🐾 Patch